requirements.txt
Django>=5.0,<5.1
Pillow
whitenoise
gunicorn
dj-database-url
psycopg2-binary
cloudinary>=1.38.0
django-cloudinary-storage>=0.3.0
django-multiselectfield
twilio
django-multiselectfield==1.0.1
python-decouple
twilio>=7.0.0
djangorestframework